hf-space/requirements.txt
rembg>=2.0.0
onnxruntime>=1.23.0
Pillow>=10.0.0
